--- app/src/main/java/cs2340/group61/doughnation/domain/AccountType.java
@@ -1,5 +1,0 @@
-package cs2340.group61.doughnation.domain;
-
-public enum AccountType {
-    ADMIN, GENERAL_USER, LOCATION_EMPLOYEE, LOCATION_MANAGER;
-}
