--- app/src/main/java/cs2340/group61/doughnation/domain/Admin.java
@@ -1,8 +1,0 @@
-package cs2340.group61.doughnation.domain;
-
-/**
- * Can add/remove users, add/remove locations, lock/unlock accounts.
- */
-public class Admin extends User {
-
-}
