--- app/src/main/java/cs2340/group61/doughnation/domain/LocationEmployee.java
@@ -1,5 +1,0 @@
-package cs2340.group61.doughnation.domain;
-
-public class LocationEmployee extends User {
-
-}
