--- app/src/main/java/cs2340/group61/doughnation/domain/Manager.java
@@ -1,4 +1,0 @@
-package cs2340.group61.doughnation.domain;
-
-public class Manager extends LocationEmployee {
-}
