--- app/src/main/java/cs2340/group61/doughnation/domain/Organization.java
@@ -1,6 +1,0 @@
-package cs2340.group61.doughnation.domain;
-
-//Sample organizations to assign managers/location employees
-public enum Organization {
-    HUMANE_SOCIETY, GOODWILL, DOCTORS_WITHOUT_BORDERS;
-}
