+++ app/src/main/java/cs2340/group61/doughnation/model/domain/AccountType.java
@@ -1,0 +1,5 @@
+package cs2340.group61.doughnation.model.domain;
+
+public enum AccountType {
+    ADMIN, GENERAL_USER, LOCATION_EMPLOYEE, ORGANIZATION_MANAGER;
+}
