+++ app/src/main/java/cs2340/group61/doughnation/model/domain/Organizations.java
@@ -1,0 +1,7 @@
+package cs2340.group61.doughnation.model.domain;
+
+//Sample organizations to assign managers/location employees
+//TODO: change Organizations enum to Organization class to be instantiated
+public enum Organizations {
+    HUMANE_SOCIETY, GOODWILL, DOCTORS_WITHOUT_BORDERS;
+}
